[BAEL-9683] Added the user_role.md created by the agent
- Improved the test name
- Used the baeldung-formatter to properly format code
