+++ spring-ai-modules/spring-ai-agent-skills/baeldung/memory/user/user_role.md
@@ -1,0 +1,6 @@
+---
+name: user_role
+description: Software engineer working on an AI article
+type: user
+---
+User is a software engineer focused on an AI article, seeking collaboration on technical content creation.
